# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Painel Financeiro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ComparisonPeriod {
  revenueChange: number;
  profitChange: number;
  expensesChange: number;
  salesChange: number;
}

export interface OverviewMetrics {
  totalRevenue: number;
  totalProfit: number;
  totalExpenses: number;
  roi: number;
  totalSales: number;
  avgTicket: number;
  totalRefunds: number;
  todayRevenue: number;
  todayProfit: number;
  todayExpenses: number;
  todaySales: number;
  period: string;
  comparisonPrevious: ComparisonPeriod;
}

export interface DailyMetric {
  date: string;
  revenue: number;
  profit: number;
  expenses: number;
  sales: number;
}

export interface PerformanceData {
  daily: DailyMetric[];
  avgRevenue: number;
  avgProfit: number;
  avgExpenses: number;
  period: string;
}

export interface Offer {
  id: string;
  name: string;
  revenue: number;
  profit: number;
  expenses: number;
  roi: number;
  sales: number;
  avgTicket: number;
  refunds: number;
}

export interface OfferDetail {
  id: string;
  name: string;
  revenue: number;
  profit: number;
  expenses: number;
  roi: number;
  sales: number;
  avgTicket: number;
  refunds: number;
  daily: DailyMetric[];
}

export interface ComparisonOffer {
  id: string;
  name: string;
  revenue: number;
  profit: number;
  roi: number;
  sales: number;
  avgTicket: number;
  daily: DailyMetric[];
}

export interface ComparisonData {
  offers: ComparisonOffer[];
  period: string;
}

export interface CashflowEntry {
  date: string;
  cumulativeRevenue: number;
  cumulativeExpenses: number;
  cumulativeProfit: number;
  dailyRevenue: number;
  dailyExpenses: number;
  dailyProfit: number;
}

export interface CashflowData {
  entries: CashflowEntry[];
  totalRevenue: number;
  totalExpenses: number;
  totalProfit: number;
  period: string;
}

export type SyncStatusStatus = typeof SyncStatusStatus[keyof typeof SyncStatusStatus];


export const SyncStatusStatus = {
  idle: 'idle',
  syncing: 'syncing',
  success: 'success',
  error: 'error',
} as const;

export interface SyncStatus {
  status: SyncStatusStatus;
  /** @nullable */
  lastSyncAt: string | null;
  offersCount: number;
  message: string;
}

export type GetOverviewParams = {
period?: GetOverviewPeriod;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type GetOverviewPeriod = typeof GetOverviewPeriod[keyof typeof GetOverviewPeriod];


export const GetOverviewPeriod = {
  today: 'today',
  yesterday: 'yesterday',
  last7days: 'last7days',
  last30days: 'last30days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;

export type GetPerformanceParams = {
period?: GetPerformancePeriod;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type GetPerformancePeriod = typeof GetPerformancePeriod[keyof typeof GetPerformancePeriod];


export const GetPerformancePeriod = {
  last7days: 'last7days',
  last30days: 'last30days',
  last90days: 'last90days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;

export type GetOffersParams = {
period?: GetOffersPeriod;
sortBy?: GetOffersSortBy;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type GetOffersPeriod = typeof GetOffersPeriod[keyof typeof GetOffersPeriod];


export const GetOffersPeriod = {
  today: 'today',
  yesterday: 'yesterday',
  last7days: 'last7days',
  last30days: 'last30days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;

export type GetOffersSortBy = typeof GetOffersSortBy[keyof typeof GetOffersSortBy];


export const GetOffersSortBy = {
  profit: 'profit',
  revenue: 'revenue',
  roi: 'roi',
  sales: 'sales',
} as const;

export type FetchOfferParams = {
period?: FetchOfferPeriod;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type FetchOfferPeriod = typeof FetchOfferPeriod[keyof typeof FetchOfferPeriod];


export const FetchOfferPeriod = {
  today: 'today',
  yesterday: 'yesterday',
  last7days: 'last7days',
  last30days: 'last30days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;

export type FetchOfferPerformanceParams = {
period?: FetchOfferPerformancePeriod;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type FetchOfferPerformancePeriod = typeof FetchOfferPerformancePeriod[keyof typeof FetchOfferPerformancePeriod];


export const FetchOfferPerformancePeriod = {
  last7days: 'last7days',
  last30days: 'last30days',
  last90days: 'last90days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;

export type GetComparisonParams = {
/**
 * Comma-separated offer IDs
 */
offerIds: string;
period?: GetComparisonPeriod;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type GetComparisonPeriod = typeof GetComparisonPeriod[keyof typeof GetComparisonPeriod];


export const GetComparisonPeriod = {
  today: 'today',
  yesterday: 'yesterday',
  last7days: 'last7days',
  last30days: 'last30days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;

export type GetCashflowParams = {
period?: GetCashflowPeriod;
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};

export type GetCashflowPeriod = typeof GetCashflowPeriod[keyof typeof GetCashflowPeriod];


export const GetCashflowPeriod = {
  last7days: 'last7days',
  last30days: 'last30days',
  last90days: 'last90days',
  thismonth: 'thismonth',
  lastmonth: 'lastmonth',
} as const;